requirements.txt
openai
langchain
pymupdf
beautifulsoup4
requests
chromadb
python-dotenv
streamlit
PyPDF2
python-docx
fastapi
uvicorn
pandas
numpy
# Legal Document Processing Pipeline
unstructured[pdf]==0.11.8
pdfminer.six==20231228
sentence-transformers==2.2.2
qdrant-client==1.7.0
langchain==0.1.0
langchain-community==0.0.10
tiktoken==0.5.2
pandas==2.1.4
numpy==1.24.3
tqdm==4.66.1
python-magic==0.4.27
python-docx==1.1.0

# Malaysian Language Processing (optional)
langdetect==1.0.9

# FastAPI for search interface
fastapi==0.104.1
uvicorn==0.24.0